refactor: 혼잡도 관련 컴포넌트 파일 위치 수정
[#363]
