+++ frontend/src/components/ui/DetailedStationInfo/congestion/CongestionStatistics.tsx
